+++ tests/testthat/test_data_selection.R
@@ -3,10 +3,7 @@ context('test data selection')
 library(dplyr)
 library(DBI)
 
-# to do: test what happens if integers or factors are used as booklet/perdon/item id's
-
-
-expect_no_error = function(object, info=NULL) expect_error(object, regexp=NA, info=info)
+# to do: test what happens if integers or factors are used as booklet/person/item id's